Add applies_to function to every policy which returns a list of PPI classes that the policy cares about (for example PPI::Document or PPI::Token::Word or PPI::Statement::Sub).  Perl::Critic::critique then collects those and only calls violates() on elements that match.
This pre-filtering is a significant performance boost.  I benchmarked
t/04_criticize.t before and after and saw a 28% time reduction.

Note: for most policies this change makes the violates() function less
fault tolerant.  Its behavior is undefined if called with a PPI
instance that is not one of the ones specified in applies_to().

Also added applies_to() to the list of functions that
Test::Pod::Coverage will ignore.